refactor test setup out
